requirements.txt
Django==2.0
PyJWT==1.5.3
chardet==2.2.1
colorama==0.2.5
command-not-found==0.3
django-cors-headers==2.1.0
djangorestframework==3.7.3
html5lib==0.999
language-selector==0.1
pbr==3.1.1
pygobject==3.12.0
python-apt==0.9.3.5ubuntu2
pytz==2017.3
requests==2.2.1
six==1.11.0
stevedore==1.27.1
ufw==0.34-rc-0ubuntu2
urllib3==1.7.1
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.8.2
wheel==0.24.0
